1、修复了pileup修正时单位不匹配的问题 2、保存数据时候 统一都改成了float类型

+++ dealusb_msg.h
@@ -159,7 +159,7 @@ public slots:
 
 
     //校正
-    int pileUp_calibration(int,int);
+    float pileUp_calibration(int,int);
 
 
     //自动校正相关
